clickhouse: simplified function formatter implementations
