+++ ibis/clickhouse/compiler.py
@@ -420,763 +420,7 @@ class ClickhouseUnion(comp.Union):
         return '\n'.join(buf)
 
 
-# ---------------------------------------------------------------------
-# Scalar and array expression formatting
-
-_sql_type_names = {
-    'int8': 'Int8',
-    'int16': 'Int16',
-    'int32': 'Int32',
-    'int64': 'Int64',
-    'float': 'Float32',
-    'double': 'Float64',
-    'string': 'String',
-    'boolean': 'UInt8',
-    'date': 'Date',
-    'timestamp': 'DateTime',
-    'decimal': 'UInt64',  # see Clickhouse issue #253
-}
-
-
-def _cast(translator, expr):
-    op = expr.op()
-    arg, target_type = op.args
-    arg_formatted = translator.translate(arg)
-
-    if isinstance(arg, ir.CategoryValue) and target_type == 'int32':
-        return arg_formatted
-    else:
-        sql_type = _type_to_sql_string(target_type)
-        return 'CAST({0!s} AS {1!s})'.format(arg_formatted, sql_type)
-
-
-def _type_to_sql_string(tval):
-    return _sql_type_names[tval.name.lower()]
-
-
-def _between(translator, expr):
-    op = expr.op()
-    comp, lower, upper = [translator.translate(x) for x in op.args]
-    return '{0!s} BETWEEN {1!s} AND {2!s}'.format(comp, lower, upper)
-
-
-_cumulative_to_reduction = {
-    ops.CumulativeSum: ops.Sum,
-    ops.CumulativeMin: ops.Min,
-    ops.CumulativeMax: ops.Max,
-    ops.CumulativeMean: ops.Mean,
-    ops.CumulativeAny: ops.Any,
-    ops.CumulativeAll: ops.All,
-}
-
-
-def _shift_like(name):
-
-    def formatter(translator, expr):
-        op = expr.op()
-        arg, offset, default = op.args
-
-        arg_formatted = translator.translate(arg)
-
-        if default is not None:
-            if offset is None:
-                offset_formatted = '1'
-            else:
-                offset_formatted = translator.translate(offset)
-
-            default_formatted = translator.translate(default)
-
-            return '{0}({1}, {2}, {3})'.format(name, arg_formatted,
-                                               offset_formatted,
-                                               default_formatted)
-        elif offset is not None:
-            offset_formatted = translator.translate(offset)
-            return '{0}({1}, {2})'.format(name, arg_formatted,
-                                          offset_formatted)
-        else:
-            return '{0}({1})'.format(name, arg_formatted)
-
-    return formatter
-
-
-def _nth_value(translator, expr):
-    op = expr.op()
-    arg, rank = op.args
-
-    arg_formatted = translator.translate(arg)
-    rank_formatted = translator.translate(rank - 1)
-
-    return 'first_value(lag({0}, {1}))'.format(arg_formatted,
-                                               rank_formatted)
-
-
-def _ntile(translator, expr):
-    op = expr.op()
-    arg, buckets = map(translator.translate, op.args)
-    return 'ntile({})'.format(buckets)
-
-
-def _negate(translator, expr):
-    arg = expr.op().args[0]
-    formatted_arg = translator.translate(arg)
-    if isinstance(expr, ir.BooleanValue):
-        return 'NOT {0!s}'.format(formatted_arg)
-    else:
-        if _needs_parens(arg):
-            formatted_arg = _parenthesize(formatted_arg)
-        return '-{0!s}'.format(formatted_arg)
-
-
-def _not(translator, expr):
-    return 'NOT {}'.format(*map(translator.translate, expr.op().args))
-
-
-def _parenthesize(what):
-    return '({0!s})'.format(what)
-
-
-def unary(func_name):
-    return fixed_arity(func_name, 1)
-
-
-# def _if_combinator(translator, func_name, arg, where):
-#     arg = translator.translate(arg)
-#     if where is not None:
-#         cond = translator.translate(where)
-#         return '{0!s}If({1!s}, {2})'.format(func_name, arg, cond)
-#     else:
-#         return '{0!s}({1!s})'.format(func_name, arg)
-
-
-def _function_call(translator, func, *args):
-    args_ = ', '.join(map(translator.translate, args))
-    return '{0!s}({1!s})'.format(func, args_)
-
-
-def _aggregate(translator, func, arg, where=None):
-    if where is not None:
-        return _function_call(translator, func + 'If', arg, where)
-    else:
-        return _function_call(translator, func, arg)
-
-
-def agg(func):
-    def formatter(translator, expr):
-        return _aggregate(translator, func, *expr.op().args)
-    return formatter
-
-
-def agg_variance_like(func):
-    variants = {'sample': '{0}Samp'.format(func),
-                'pop': '{0}Pop'.format(func)}
-
-    def formatter(translator, expr):
-        arg, where, how = expr.op().args
-        return _aggregate(translator, variants[how], arg, where)
-
-    return formatter
-
-
-def fixed_arity(func_name, arity):
-    def formatter(translator, expr):
-        op = expr.op()
-        arg_count = len(op.args)
-        if arity != arg_count:
-            msg = 'Incorrect number of args {0} instead of {1}'
-            raise com.IbisError(msg.format(arg_count, arity))
-        return _function_call(translator, func_name, *op.args)
-    return formatter
-
-
-def _binary_infix_op(infix_sym):
-    def formatter(translator, expr):
-        op = expr.op()
-
-        left, right = op.args
-
-        left_arg = translator.translate(left)
-        right_arg = translator.translate(right)
-
-        if _needs_parens(left):
-            left_arg = _parenthesize(left_arg)
-
-        if _needs_parens(right):
-            right_arg = _parenthesize(right_arg)
-
-        return '{0!s} {1!s} {2!s}'.format(left_arg, infix_sym, right_arg)
-    return formatter
-
-
-def _xor(translator, expr):
-    op = expr.op()
-
-    left_arg = translator.translate(op.left)
-    right_arg = translator.translate(op.right)
-
-    if _needs_parens(op.left):
-        left_arg = _parenthesize(left_arg)
-
-    if _needs_parens(op.right):
-        right_arg = _parenthesize(right_arg)
-
-    return 'xor({0}, {1})'.format(left_arg, right_arg)
-
-
-def _name_expr(formatted_expr, quoted_name):
-    return '{0!s} AS {1!s}'.format(formatted_expr, quoted_name)
-
-
-def _needs_parens(op):
-    if isinstance(op, ir.Expr):
-        op = op.op()
-    op_klass = type(op)
-    # function calls don't need parens
-    return (op_klass in _binary_infix_ops or
-            op_klass in [ops.Negate])
-
-
-def _need_parenthesize_args(op):
-    if isinstance(op, ir.Expr):
-        op = op.op()
-    op_klass = type(op)
-    return (op_klass in _binary_infix_ops or
-            op_klass in [ops.Negate])
-
-
-def _boolean_literal(expr):
-    value = expr.op().value
-    return '1' if value else '0'
-
-
-def _number_literal(expr):
-    value = expr.op().value
-    return repr(value)
-
-
-def _string_literal(expr):
-    value = expr.op().value
-    return "'{0!s}'".format(value.replace("'", "\\'"))
-
-
-def _date_literal(expr):
-    value = expr.op().value
-    if isinstance(value, datetime.date):
-        value = value.strftime('%Y-%m-%d')
-    return "toDate('{0!s}')".format(value)
-
-
-def _timestamp_literal(expr):
-    value = expr.op().value
-    if isinstance(value, datetime.datetime):
-        if value.microsecond != 0:
-            raise ValueError('Unspoorted subsecond accuracy {}'.format(value))
-        value = value.strftime('%Y-%m-%d %H:%M:%S')
-    return "toDateTime('{0!s}')".format(value)
-
-
-# def _null_literal_format(translator, expr):
-#     return 'NULL'
-
-
-def _literal(translator, expr):
-    if isinstance(expr, ir.BooleanValue):
-        return _boolean_literal(expr)
-    elif isinstance(expr, ir.StringValue):
-        return _string_literal(expr)
-    elif isinstance(expr, ir.NumericValue):
-        return _number_literal(expr)
-    elif isinstance(expr, ir.TimestampValue):
-        return _timestamp_literal(expr)
-    elif isinstance(expr, ir.DateValue):
-        return _date_literal(expr)
-    else:
-        raise NotImplementedError
-
-
-class CaseFormatter(object):
-
-    def __init__(self, translator, base, cases, results, default):
-        self.translator = translator
-        self.base = base
-        self.cases = cases
-        self.results = results
-
-        if default.equals(ir.null()):
-            raise TypeError('Null is not supported, default '
-                            'value must be explicitly defined!')
-        self.default = default
-
-        # HACK
-        self.indent = 2
-        self.multiline = len(cases) > 1
-        self.buf = StringIO()
-
-    def _trans(self, expr):
-        return self.translator.translate(expr)
-
-    def get_result(self):
-        self.buf.seek(0)
-
-        self.buf.write('CASE')
-        if self.base is not None:
-            base_str = self._trans(self.base)
-            self.buf.write(' {0}'.format(base_str))
-
-        for case, result in zip(self.cases, self.results):
-            self._next_case()
-            case_str = self._trans(case)
-            result_str = self._trans(result)
-            self.buf.write('WHEN {0} THEN {1}'.format(case_str, result_str))
-
-        if self.default is not None:
-            self._next_case()
-            default_str = self._trans(self.default)
-            self.buf.write('ELSE {0}'.format(default_str))
-
-        if self.multiline:
-            self.buf.write('\nEND')
-        else:
-            self.buf.write(' END')
-
-        return self.buf.getvalue()
-
-    def _next_case(self):
-        if self.multiline:
-            self.buf.write('\n{0}'.format(' ' * self.indent))
-        else:
-            self.buf.write(' ')
-
-
-def _simple_case(translator, expr):
-    op = expr.op()
-    formatter = CaseFormatter(translator, op.base, op.cases, op.results,
-                              op.default)
-    return formatter.get_result()
-
-
-def _searched_case(translator, expr):
-    op = expr.op()
-    formatter = CaseFormatter(translator, None, op.cases, op.results,
-                              op.default)
-    return formatter.get_result()
-
-
-def _table_array_view(translator, expr):
-    ctx = translator.context
-    table = expr.op().table
-    query = ctx.get_compiled_expr(table)
-    return '(\n{0}\n)'.format(util.indent(query, ctx.indent))
-
-
-# ---------------------------------------------------------------------
-# Timestamp arithmetic and other functions
-
-def _timestamp_delta(translator, expr):
-    op = expr.op()
-    arg, offset = op.args
-    formatted_arg = translator.translate(arg)
-    return _timestamp_format_offset(offset, formatted_arg)
-
-
-_clickhouse_delta_functions = {
-    tempo.Year: 'years_add',
-    tempo.Month: 'months_add',
-    tempo.Week: 'weeks_add',
-    tempo.Day: 'days_add',
-    tempo.Hour: 'hours_add',
-    tempo.Minute: 'minutes_add',
-    tempo.Second: 'seconds_add',
-    tempo.Millisecond: 'milliseconds_add',
-    tempo.Microsecond: 'microseconds_add',
-    tempo.Nanosecond: 'nanoseconds_add'
-}
-
-
-def _timestamp_format_offset(offset, arg):
-    f = _clickhouse_delta_functions[type(offset)]
-    return '{0}({1}, {2})'.format(f, arg, offset.n)
-
-
-# ---------------------------------------------------------------------
-# Semi/anti-join supports
-
-
-def _exists_subquery(translator, expr):
-    op = expr.op()
-    ctx = translator.context
-
-    dummy = ir.literal(1).name(ir.unnamed)
-
-    filtered = op.foreign_table.filter(op.predicates)
-    expr = filtered.projection([dummy])
-
-    subquery = ctx.get_compiled_expr(expr)
-
-    if isinstance(op, transforms.ExistsSubquery):
-        key = 'EXISTS'
-    elif isinstance(op, transforms.NotExistsSubquery):
-        key = 'NOT EXISTS'
-    else:
-        raise NotImplementedError
-
-    return '{0} (\n{1}\n)'.format(key, util.indent(subquery, ctx.indent))
-
-
-def _table_column(translator, expr):
-    op = expr.op()
-    field_name = op.name
-    quoted_name = quote_identifier(field_name, force=True)
-
-    table = op.table
-    ctx = translator.context
-
-    # If the column does not originate from the table set in the current SELECT
-    # context, we should format as a subquery
-    if translator.permit_subquery and ctx.is_foreign_expr(table):
-        proj_expr = table.projection([field_name]).to_array()
-        return _table_array_view(translator, proj_expr)
-
-    if ctx.need_aliases():
-        alias = ctx.get_ref(table)
-        if alias is not None:
-            quoted_name = '{0}.{1}'.format(alias, quoted_name)
-
-    return quoted_name
-
-
-def _timestamp_from_unix(translator, expr):
-    op = expr.op()
-
-    val, unit = op.args
-
-    if unit == 'ms':
-        raise ValueError('`ms` unit is not supported!')
-    elif unit == 'us':
-        raise ValueError('`us` unit is not supported!')
-
-    arg = translator.translate(val)
-    return 'toUInt32({0})'.format(arg)
-
-
-def varargs(func_name):
-    def varargs_formatter(translator, expr):
-        op = expr.op()
-        return _function_call(translator, func_name, *op.args)
-    return varargs_formatter
-
-
-def _substring(translator, expr):
-    # arg_ is the formatted notation
-    op = expr.op()
-    arg, start, length = op.args
-    arg_, start_ = translator.translate(arg), translator.translate(start)
-
-    # Clickhouse is 1-indexed
-    if length is None or isinstance(length.op(), ir.Literal):
-        if length is not None:
-            length_ = length.op().value
-            return 'substring({0}, {1} + 1, {2})'.format(arg_, start_, length_)
-        else:
-            return 'substring({0}, {1} + 1)'.format(arg_, start_)
-    else:
-        length_ = translator.translate(length)
-        return 'substring({0}, {1} + 1, {2})'.format(arg_, start_, length_)
-
-
-def _string_find(translator, expr):
-    op = expr.op()
-    arg, substr, start, _ = op.args
-    if start is not None:
-        raise com.IbisError('String find doesn\'t support start argument')
-
-    arg_, substr_ = translator.translate(arg), translator.translate(substr)
-    return 'position({0}, {1}) - 1'.format(arg_, substr_)
-
-
-def _regex_extract(translator, expr):
-    op = expr.op()
-    arg, pattern, index = op.args
-    arg_, pattern_ = translator.translate(arg), translator.translate(pattern)
-
-    if index is not None:
-        index_ = translator.translate(index)
-        return 'extractAll({0}, {1})[{2} + 1]'.format(arg_, pattern_, index_)
-
-    return 'extractAll({0}, {1})'.format(arg_, pattern_)
-
-
-def _string_join(translator, expr):
-    op = expr.op()
-    arg, strings = op.args
-    return _function_call(translator, 'concat_ws', arg, *strings)
-
-
-def _parse_url(translator, expr):
-    op = expr.op()
-
-    arg, extract, key = op.args
-    arg_formatted = translator.translate(arg)
-
-    if key is None:
-        return "parse_url({0}, '{1}')".format(arg_formatted, extract)
-    else:
-        key_fmt = translator.translate(key)
-        return "parse_url({0}, '{1}', {2})".format(arg_formatted,
-                                                   extract, key_fmt)
-
-
-def _index_of(translator, expr):
-    op = expr.op()
-
-    arg, arr = op.args
-    arg_formatted = translator.translate(arg)
-    arr_formatted = ','.join(map(translator.translate, arr))
-    return "indexOf([{0}], {1}) - 1".format(arr_formatted, arg_formatted)
-
-
-def _sign(translator, expr):
-    """Workaround for missing sign function"""
-    op = expr.op()
-    arg, = op.args
-    arg_ = translator.translate(arg)
-    return 'intDivOrZero({0}, abs({0}))'.format(arg_)
-
-
-def _round(translator, expr):
-    op = expr.op()
-    arg, digits = op.args
-
-    arg_ = translator.translate(arg)
-    if digits is not None:
-        digits_ = translator.translate(digits)
-        return 'round({0}, {1})'.format(arg_, digits_)
-    else:
-        return 'round({0})'.format(arg_)
-
-
-# TODO there are a lot of hash functions in clickhouse
-def _hash(translator, expr):
-    op = expr.op()
-    arg, how = op.args
-
-    arg_formatted = translator.translate(arg)
-
-    if how == 'fnv':
-        return 'fnv_hash({0})'.format(arg_formatted)
-    else:
-        raise NotImplementedError(how)
-
-
-def _log(translator, expr):
-    op = expr.op()
-    arg, base = op.args
-    arg_formatted = translator.translate(arg)
-
-    # I don't know how to check base value properly
-    if base is None:
-        return 'log({0})'.format(arg_formatted)
-    elif base._arg.value == 2:
-        return 'log2({0})'.format(arg_formatted)
-    elif base._arg.value == 10:
-        return 'log10({0})'.format(arg_formatted)
-    else:
-        raise ValueError('Base {} for logarithm not supported!'.format(base))
-
-
-def _count_distinct(translator, expr):
-    # clickhouse supports both COUNT(DISTINCT x) and uniq()
-    op = expr.op()
-    arg_formatted = translator.translate(op.args[0])
-    return 'COUNT(DISTINCT {0})'.format(arg_formatted)
-
-
-def _value_list(translator, expr):
-    op = expr.op()
-    formatted = [translator.translate(x) for x in op.values]
-    return '({0})'.format(', '.join(formatted))
-
-
-def _identical_to(translator, expr):
-    op = expr.op()
-    if op.args[0].equals(op.args[1]):
-        return '1'
-
-    left, right = map(translator.translate, op.args)
-    return '{left} IS NOT DISTINCT FROM {right}'.format(
-        left=left, right=right
-    )
-
-
-_subtract_one = '({0} - 1)'.format
-
-
-_expr_transforms = {
-    ops.RowNumber: _subtract_one,
-    ops.DenseRank: _subtract_one,
-    ops.MinRank: _subtract_one,
-    ops.NTile: _subtract_one,
-}
-
-
-# TODO: clickhouse uses differenct string functions
-#       for ascii and utf-8 encodings,
-
-_binary_infix_ops = {
-    # Binary operations
-    ops.Add: _binary_infix_op('+'),
-    ops.Subtract: _binary_infix_op('-'),
-    ops.Multiply: _binary_infix_op('*'),
-    ops.Divide: _binary_infix_op('/'),
-    ops.Power: fixed_arity('pow', 2),
-    ops.Modulus: _binary_infix_op('%'),
-
-    # Comparisons
-    ops.Equals: _binary_infix_op('='),
-    ops.NotEquals: _binary_infix_op('!='),
-    ops.GreaterEqual: _binary_infix_op('>='),
-    ops.Greater: _binary_infix_op('>'),
-    ops.LessEqual: _binary_infix_op('<='),
-    ops.Less: _binary_infix_op('<'),
-    ops.IdenticalTo: _identical_to,
-
-    # Boolean comparisons
-    ops.And: _binary_infix_op('AND'),
-    ops.Or: _binary_infix_op('OR'),
-    ops.Xor: _xor,
-}
-
-
-_operation_registry = {
-    # Unary operations
-    ops.TypeOf: unary('toTypeName'),
-    # ops.NotNull: _not_null,
-    # ops.IsNull: _is_null,
-    ops.Negate: _negate,
-    ops.Not: _not,
-
-    # ops.IfNull: _ifnull_workaround,
-    # ops.NullIf: fixed_arity('nullif', 2),
-
-    # ops.ZeroIfNull: unary('zeroifnull'),
-    # ops.NullIfZero: unary('nullifzero'),
-
-    ops.Abs: unary('abs'),
-    # ops.BaseConvert: fixed_arity('conv', 3),
-    ops.Ceil: unary('ceil'),
-    ops.Floor: unary('floor'),
-    ops.Exp: unary('exp'),
-    ops.Round: _round,
-
-    ops.Sign: _sign,
-    ops.Sqrt: unary('sqrt'),
-
-    ops.Hash: _hash,
-
-    ops.Log: _log,
-    ops.Ln: unary('log'),
-    ops.Log2: unary('log2'),
-    ops.Log10: unary('log10'),
-
-    # ops.DecimalPrecision: unary('precision'),
-    # ops.DecimalScale: unary('scale'),
-
-    # Unary aggregates
-    ops.CMSMedian: agg('median'),
-    # TODO: there is also a `uniq` function which is the
-    #       recommended way to approximate cardinality
-    ops.HLLCardinality: agg('uniqHLL12'),
-    ops.Mean: agg('avg'),
-    ops.Sum: agg('sum'),
-    ops.Max: agg('max'),
-    ops.Min: agg('min'),
-
-    ops.StandardDev: agg_variance_like('stddev'),
-    ops.Variance: agg_variance_like('var'),
-
-    # ops.GroupConcat: fixed_arity('group_concat', 2),
-
-    ops.Count: agg('count'),
-    ops.CountDistinct: agg('uniq'), # _count_distinct,
-
-    # string operations
-    ops.StringLength: unary('length'),
-    ops.Lowercase: unary('lower'),
-    ops.Uppercase: unary('upper'),
-    ops.Reverse: unary('reverse'),
-    ops.Substring: _substring,
-    ops.StringFind: _string_find,
-    ops.FindInSet: _index_of,
-    ops.StringReplace: fixed_arity('replaceAll', 3),
-
-    # TODO: there are no concat_ws in clickhouse
-    # ops.StringJoin: varargs('concat'),
-
-    ops.StringSQLLike: _binary_infix_op('LIKE'),
-    ops.RegexSearch: fixed_arity('match', 2),
-    # TODO: extractAll(haystack, pattern)[index + 1]
-    ops.RegexExtract: _regex_extract,
-    ops.RegexReplace: fixed_arity('replaceRegexpAll', 3),
-    ops.ParseURL: _parse_url,
-
-    # Timestamp operations
-    ops.TimestampNow: lambda *args: 'now()',
-    ops.ExtractYear: unary('toYear'),
-    ops.ExtractMonth: unary('toMonth'),
-    ops.ExtractDay: unary('toDayOfMonth'),
-    ops.ExtractHour: unary('toHour'),
-    ops.ExtractMinute: unary('toMinute'),
-    ops.ExtractSecond: unary('toSecond'),
-    # ops.Truncate: _truncate,
-
-    # Other operations
-    ops.E: lambda *args: 'e()',
-
-    ir.Literal: _literal,
-    # ir.NullLiteral: _null_literal,
-
-    ir.ValueList: _value_list,
-
-    ops.Cast: _cast,
-
-    ops.Coalesce: varargs('coalesce'),
-    ops.Greatest: varargs('greatest'),
-    ops.Least: varargs('least'),
-
-    ops.Where: fixed_arity('if', 3),
-
-    ops.Between: _between,
-    ops.Contains: _binary_infix_op('IN'),
-    ops.NotContains: _binary_infix_op('NOT IN'),
-
-    ops.SimpleCase: _simple_case,
-    ops.SearchedCase: _searched_case,
-
-    ops.TableColumn: _table_column,
-
-    ops.TableArrayView: _table_array_view,
-
-    ops.TimestampDelta: _timestamp_delta,
-    ops.TimestampFromUNIX: _timestamp_from_unix,
-
-    transforms.ExistsSubquery: _exists_subquery,
-    transforms.NotExistsSubquery: _exists_subquery,
-
-    # RowNumber, and rank functions starts with 0 in Ibis-land
-    # ops.RowNumber: lambda *args: 'row_number()',
-    # ops.DenseRank: lambda *args: 'dense_rank()',
-    # ops.MinRank: lambda *args: 'rank()',
-    # ops.PercentRank: lambda *args: 'percent_rank()',
-
-    ops.FirstValue: unary('first_value'),
-    ops.LastValue: unary('last_value'),
-    ops.NthValue: _nth_value,
-    ops.Lag: _shift_like('lag'),
-    ops.Lead: _shift_like('lead'),
-    # ops.WindowOp: _window,
-    ops.NTile: _ntile,
-}
-
-_operation_registry.update(_binary_infix_ops)
+from .operations import _operation_registry, _name_expr, _type_to_sql_string
 
 
 class ClickhouseExprTranslator(comp.ExprTranslator):
